Init with QEMU testing and decoder
Testing:
For testing and to emulate the later usage as a bare metal library we use QEMU to run RISC-V code.
This should also help iron out an endianess issues which might happen.

Decoder:
We will split the library into two phases: decoding and tracing.
The decoder will decompress (if necessary) and parse the output of the decoder
and produce easily usable Rust structs or enums.
The tracer will use those to do the actual tracing. While this might not the most performant
design, it will allow easy re-use of components and introspective into the entire tracing flow.