Split up tracing error type
We always want to return the tracer state if an error has occurred, but we also want to easily change the actual error type. We split TraceError into TraceError which will hold the state and the underlying error with TraceErrorType. We can thus easily add new errors, without having to explicitly add the tracing state to each.
Also add proper debug printing.