Pass packet data as slice into Decoder::read
Before, the Decoder held the packet data as a slice wrapped in an Option as for the duration of a read call. This complicates the lifetime of the Decoder and the user needs to set the packet buffer length by hand. By just passing a slice, we have the same behavior as before but don't emulate function parameters in a convoluted way and also remove the amount of possible bugs. Users no longer need to set the buffer.