Use memory slice in segments
To return this crate into the lands of safe Rust we replace the unsafe pointer dereference and complicated calculation in the disassembler. Instead, a slice can be used in the segments and disasm, which is both safe and easier to use. This also prevents further accidents and misuse in the future.