Add instruction cache
Many times we are reading the same instructions in the same basic block over and over again. We do not need to decode the same instruction multiple times.

We introduce a preliminary and plain LRU cache for instructions. In the future a better, more basic block aware cache might yield even better performance.