Add feature to select v1 or v2 tracing
The version 1 tracing algorithm can be turned on with the "tracing_v1" feature flag.
It should only be used for testing purposes as it misses some pc as reported in https://github.com/riscv-non-isa/riscv-trace-spec/issues/89.
This is the only change in the decoder algorithm between v1 and v2.

The current change of algorithm behaviour with feature flags is not really future prove and should be revisited if future versions of the tracing algorithm are to be supported in parallel.