tracer: remove TraceError
The current `TraceError` was introduced a while back. It bundles a
`TraceErrorType` with the current `TraceState`. While that state _may_
be of interest to some users, it's also quite large. We therefore deem
it unfit for inclusion as an error type. Since `TraceError` is nothing
but a struct encapsulating `TraceState` and `TraceErrorType`, we decided
to remove it.