instruction: allow extraction of instruction type data from compressed
Currently, `Instruction` has a few fields which may or may not be set
depending on what instruction was decoded. And which instruction was
decoded can be gathered from the `Instruction::name` if it's not `None`.
This leads to some convoluted logic in both the decoding and query fns.

We thus decided to include more of that possibly relevant information
in the `Name` type. In particular, this includes registers involved and
immediates.

We already introduced types for RISC-V 32bit instruction encodings.
However, we'll also need to handle compressed instructions. Since all of
them expand to 32bit instructions, we decided to use types we already
have and simply add new `From<u16>` impls.