instruction: add fn to determine a Kind's branch target
Currently, `Instruction` has a few fields which may or may not be set
depending on what instruction was decoded. And which instruction was
decoded can be gathered from the `Instruction::kind` if it's not `None`.
This leads to some convoluted logic in both the decoding and query fns.

We thus include more of that possibly relevant information in the `Kind`
type, although we currently make little use of that. We could now
rewrite `Instruction`s query fns based on that data. However, it seems
more appropriate to make those query fns associated with `Kind` rather
than having a `if let Some(...) = self.kind` in every single query fn.

This change introduces a fn for querying branch targets. In time, it
will replace manual queries of `Instruction::is_branch` and `imm`.