instruction: make Name carry instruction data
Currently, `Instruction` has a few fields which may or may not be set
depending on what instruction was decoded. And which instruction was
decoded can be gathered from the `Instruction::name` if it's not `None`.
This leads to some convoluted logic in both the decoding and query fns.

We thus decided to include more of that possibly relevant information
in the `Name` type. In particular, this includes registers involved and
immediates. For this purpose, we already introduced types reflecting the
standard RISC-V instruction types. This change now makes `Name` variants
the appropriate data.