instruction: add fn for check whether a Kind is an uninferable discon
Currently, `Instruction` has a few fields which may or may not be set
depending on what instruction was decoded. And which instruction was
decoded can be gathered from the `Instruction::kind` if it's not `None`.
This leads to some convoluted logic in both the decoding and query fns.

We thus include more of that possibly relevant information in the `Kind`
type, although we currently make little use of that. We could now
rewrite `Instruction`s query fns based on that data. However, it seems
more appropriate to make those query fns associated with `Kind` rather
than having a `if let Some(...) = self.kind` in every single query fn.

This change introduces a fn for determining whether an instruction
causes an uninferable discontinuity. In time, it will replace the
corresponding `Instruction`. Note that we have only one variant for both
`tracing_v1` and not `tracing_v1`. We determined that the logic is, in
the end, equivalent for both 1.1.3-Frozen and 2.0.3 of the Efficient
Trace for RISC-V spec.