instruction: repurpose Instruction tests for Kind
We are in the process of moving most interesting information, and the
query fns associated with them, from `Instruction` to `Kind`. In the
future, we will also remove quite a few members from `Instruction`,
making the current tests not compile any more. Still, the tests are
useful, and since the inputs and outputs are largely the same, we can
recycle them for `Kind`.