tracer: add an InstructionCache trait in a dedicated submodule
Currently, we compile and include `disassembler::InstructionCache` in
`TraceState` only if we build with the `cache` feature. While we do want
to avoid runtime variability if possible, using features was maybe not
the right choice. It makes some code more awkward and prevents users
from making caching runtime configurable if desired.

We threfore decided to switch to another approach and abstract caches
via a trait. If we do want a cache, we supply a working one. If we
don't, we supply a dummy cache that stores nochting and returns nothing.
This change introduces just the trait. And since the cache doesn't have
anything to do with dissasembly (aside from it caching `Instruction`s)
we put it into a new, dedicated submodule.