tracer: migrate TraceState to new InstructionCaches
Previously, we compiled and included `disassembler::InstructionCache` in
`TraceState` only if we build with the `cache` feature. While we do want
to avoid runtime variability if possible, using features was maybe not
the right choice. It makes some code more awkward and prevents users
from making caching runtime configurable if desired.

We threfore decided to switch to another approach and abstract caches
via a trait. If we do want a cache, we supply a working one. If we
don't, we supply a dummy cache that stores nochting and returns nothing.

This change makes `TraceState` and `Tracer` generic over the cache
implementation, allowing users to supply one or the other.