tracer: rename TraceErrorType to plain Error
`TraceErrorType` is a mouthful. The `Trace` bit is redundant since it
resides in the `tracer` module. We much prefer writing `tracer::Smth`
over imports of entities with such a redundant name. And the `Type` bit
was motivated by the (now obsolete and removed) old `TraceError` and
thus no longer relevant. It would be, if we intended to use this enum as
an error Kind, but for now we intend to use it as an error type
directly.