tracer: add InstructionCache impl for fixed sized cache
Currently, we compile and include `disassembler::InstructionCache` in
`TraceState` only if we build with the `cache` feature. We decided to
switch to another approach for compile time variability and abstract
caches via a trait.

We could impl the new `cache::InstructionCache` trait for the existing
`disassembler::InstructionCache`. But we want to move it to a different
location anyway. And the existing implementation does a lot of pointless
copying. We threfore decided to supply a new implementation that will,
in time, replace the old one.