tracer: add a dummy InstructionCache
Currently, we compile and include `disassembler::InstructionCache` in
`TraceState` only if we build with the `cache` feature. We decided to
switch to another approach for compile time variability and abstract
caches via a trait.

This change introduces an implementation that we may use if we don't
actually want a cache.