decoder: make Decoder carry data
Currently, `Decode::decode` and some utility fns, including `Decoder`
fns, take the complete data as arguments in addition to the `Decoder`.
Considering that a given `Decoder` needs to be used with the exact same
data for proper decoding, that's bad design. A `Decoder` should carry
the data its used to decode with it.

This change makes `Decoder` carry `data`, though it's never initialized
to anything but an empty slice. Future changes will add the necessary
initialization and work towards elimination of extra `slice` parameters
in various places.