decoder: impl Clone for Decoder
We plan to make `Decoder` encapsulate the data that's being decoded, not
only the current bit position within that slice of bytes. This changes
how we interact with a `Decoder` as a user.

It may appear strange to impl `Clone` for `Decoder` now, if we plan for
it to encapsulate more data. However, the ability to clone a `Decoder`
will at least help transitioning `Decoder::decode_packet`. And long
term, we deem it acceptable.