decoder: add fn for setting (new) data to decode
Currently, `Decode::decode` and some utility fns, including `Decoder`
fns, take the complete data as arguments in addition to the `Decoder`.
Considering that a given `Decoder` needs to be used with the exact same
data for proper decoding, that's bad design.

We thus recently introduced a `data` member. However, we did not include
any means to initialize it until now.