decoder: create new decoders in Decoder::decode_packet
Currently, `Decode::decode` and some utility fns, including `Decoder`
fns, take the complete data as arguments in addition to the `Decoder`.
Considering that a given `Decoder` needs to be used with the exact same
data for proper decoding, that's bad design. We thus recently introduced
a `data` member which will be used as a data source in the future.

However, we don't want any tests failing during the transition period.
To make the transition smooth, we initialize the data using the new fn
`with_data`, which —contrary to a `reset`— does create a new `Decoder`.