decoder: add new Error variant for insufficient data
Currently, `Decode::decode` and some utility fns, including `Decoder`
fns, take the complete data as arguments in addition to the `Decoder`.
Considering that a given `Decoder` needs to be used with the exact same
data for proper decoding, that's bad design. We thus recently introduced
a `data` member which will be used as a data source in the future.

We furthermore plan to make a `Decoder` always decompress, and to embed
this in the retrieval of individual bytes rather than using a buffer
with a complete packet. This eliminates most instances where we create
an `Error::ReadTooLong`. However, we'll still need to communicate
failure to decode, e.g. if we don't have any data at all that we could
pull the sign to expand from. Still, the emantics of `ReadTooLong` don't
fit there.