decoder: add new internal Decoder fn for retrieving one byte
Currently, `Decode::decode` and some utility fns, including `Decoder`
fns, take the complete data as arguments in addition to the `Decoder`.
Considering that a given `Decoder` needs to be used with the exact same
data for proper decoding, that's bad design. We thus recently introduced
a `data` member which will be used as a data source in the future.

We furthermore want to make a `Decoder` always decompress, since
compression being optional is more motivated by debugging the sending
side rather than the receiver (i.e. us). And we want to avoid having
potentially large buffers of some fixed size just for decompression if
we can.

This change prepares the migration/replacement of `Decoder::read` and
`Decoder::read_bits` with a fn that takes care of transparently
decompressing `data`.