decoder: rewrite Decoder::read_bit so it doesn't take a slice
Currently, `Decode::decode` and some utility fns, including `Decoder`
fns, take the complete data as arguments in addition to the `Decoder`.
Considering that a given `Decoder` needs to be used with the exact same
data for proper decoding, that's bad design. We thus recently introduced
a `data` member which will be used as a data source in the future.

This change makes `Decoder::read_bit` not take a slice and decompress
transparently.