decoder: add Decoder fn for reading a set number of bits
Currently, `Decode::decode` and some utility fns, including `Decoder`
fns, take the complete data as arguments in addition to the `Decoder`.
Considering that a given `Decoder` needs to be used with the exact same
data for proper decoding, that's bad design. We thus recently introduced
a `data` member which will be used as a data source in the future.

This change introduces a replacement for `Decoder::read` that does not
take a slice, decompresses transparently and is generic over some
integer types.