lib: use u8 for bit widths in ProtocolConfiguration
`ProtocolConfiguration` contains quite a few values, most of which are
withs of fields in number of bits. Previously, those would be of type
`usize`. Given that we can currently extract at most 64 bits at a time,
this is clearly overkill.

We are reasonably sure that we will not encounter bitwidths that exceed
`64` in the near term, maybe `128` in the long term. An `u8` is enough
to represent those.