decoder: add utility trait for truncating integer types
Currently, `Decode::decode` and some utility fns, including `Decoder`
fns, take the complete data as arguments in addition to the `Decoder`.
Considering that a given `Decoder` needs to be used with the exact same
data for proper decoding, that's bad design. We thus recently introduced
a `data` member which will be used as a data source in the future.

This involves replacing the current `Decoder::read`. Given the
opportunity, we want the replacement to be more generic, i.e. capable of
not just returning an `u64` but other integer types. And ideally without
just casting the result. In particular, we want unsigned and signed
values to be expanded in a sensible way.

Extracting bytes and assembling them can already be done with the traits
`core::ops` provides. However, the sensible expansion of a given value
is hard: there is no easy way, with `core` utilities, to detect whether
an integer type is signed or unsigned. Neither do we have a generic
`checked_shl` or, alternatively, a reliable bitwidth (and we will _not_
compute them on run-time with shifts).

Thus, we sadly need some code-duplication. Moving that into a trait
makes it more manageable and contained.