decoder: rewrite payload::read_address with Decoder::read_bits
Currently, `Decode::decode` and some utility fns take the complete data
as arguments in addition to the `Decoder`. Considering that a given
`Decoder` needs to be used with the exact same data for proper decoding,
that's bad design.

We thus recently introduced a `data` member which will be used as a data
source as well as `Decoder` fns that extract data based on that. This
change now migrates `payload::read_address` over to the new extraction
fns and thus eliminates the need for its `slice` parameter.