decoder: make Decode::decode take only a Decoder, not a slice
Previously, `Decode::decode` would take the complete data to decode as
argument in addition to the `Decoder`. Considering that a given
`Decoder` needs to be used with the exact same data for proper decoding,
that's bad design.

We thus introduced the member `Decoder::data` and made all its
extraction fns extract from that data. This change now finally
eliminated the explicit data parameter from `Decoder::decode`.