decoder: remove DecoderConfiguration
We want to make a `Decoder` always decompress, since compression being
optional is more motivated by debugging the sending side rather than the
receiver (i.e. us). And we replaced the existing decompression logic in
`Decoder` with transparent decompression.

Since the `DecoderConfiguraiton` contains nothing but the decompression
setting, we decided to remove it alltogether.