decoder: add Decoder fn for reading a differential bit
Some payloads facilitate the particular compression used by the tracing
format by encoding single-bit information differentially: the actual
information is encoded in the value of a bit relative to the previous
one.

Currently, we burden consumers of those payloads to figure this out.
Given that it's trivial to do while decoding we decided to not decode
the payloads completely bare but instead convey the semantics. This
change introduces a fn that allows extracting single differential bits.
Future changes will adapt payloads that carry differentially encoded
information.