decoder: make AddressInfo::notify and updiscon differential
Some payloads facilitate the particular compression used by the tracing
format by encoding single-bit information differentially: the actual
information is encoded in the value of a bit relative to the previous
one.

Previously, we would burden consumers of this payloads to figure this
out. Given that it's trivial to do while decoding we decided to not
decode the payloads completely bare but instead convey the semantics.