decoder: make Decoder::decode_packet leave Decoder in sane state
Previously, `Decode::decode` would take the complete data to decode as
argument in addition to the `Decoder`. Considering that a given
`Decoder` needs to be used with the exact same data for proper decoding,
that's bad design. We thus introduced the member `Decoder::data` and
made all its fns, including `Decoder::decode_packet`, extract from that
data.

Historically a `Decoder` would need to be reset before decoding a packet
since we would regularly leave the `Decoder` in an arbitrary state. This
was done in `Decoder::decode_packet` when it still took a slice.

This situation is obviously not nice. Thus, we decided to take a little
extra effort and leave the `Decoder` in a state that allows decoding the
next packet as long as the underlying data does contain a complete
packet.