tracer: repurpose is_branch_taken by making it return target
Originally, `Tracer::next_pc` and many other `Tracer` fns followed the
reference pseudo-code given in the specification. This lead to deeply
unideomatic Rust code. The pseudo-code of `Tracer::next` specifically
mainly consists of checks, often matching against some enum variant, and
gated by that the logic for computing the next PC that, in our code,
need to replicate much of the work done by the previous check.

This change makes use of `instruction::Kind::branch_target` returning an
`Option`. With that, we rewrite parts of `Tracer::is_branch_taken` and
make it also return an `Option`. Since the fn now does not only perform
a check we decided to rename it.