instruction: rename Name to Kind
Currently, `Instruction` has a few fields which may or may not be set
depending on what instruction was decoded. And which instruction was
decoded can be gathered from the `Instruction::name` if it's not `None`.
This leads to some convoluted logic in both the decoding and query fns.

For this reason, we make `Name` include more of that possibly relevant
information. This, however, made its old name unsuitable. Rather than a
mere name, values of this type now denote one of several known
instruction kinds. We thus decided to rename the type to `Kind`, and to
also rename the `Instruction` member. We will not completely eliminate
the old `name` from function signatures and bodies for now. However, we
expect the name to vanish over time.