instruction: remove old Instruction query fns
These members of `Instruction` were computed by a convoluted process. In
order to eliminate this logic, we emancipated the former `Name` enum,
which is now `Kind`. We then migrated all users of the old legacy fns
and queries to members other than `kind` and `size`.

This change now removes the legacy query fns.