lib: always include all values in ProtocolConfiguration
Previously, we would include some values only with the `implicit_return`
feature. However, we want to move away from features as means for
compile time variability for protocol features and use trait-base
approaches instead. At the same time, we want to preserve the flat
`ProtocolConfiguration` structure as it reflects Table 40 of the spec.

We therefore decided to just always include those values and, in the
future, make creating a `Tracer` error if the configuration demands a
feature that is not implemented by the compile-time configuration.

Because the default was, historically, not having the feature, and
because we don't want to throw those future errors with default
configurations, we choose new defaults that are equivalent to the
feature not being active.