decoder: add fn for reading implicit return info as an Option
Currently, the `irreport` and `irdepth` fields defined by the
specification for `AddressInfo` and `JumpTargetIndex` payloads are
encapsulated in the `ImplicitReturn` that is only included and decoded
with the `implicit_return` feature. We plan to avoid using features for
compile time variance regarding protocol features. But rather than just
removing the `#[cfg()]` directives, we decided to replace that struct
entirely with a simple `Option` that captures the semantics of that
struct.

This change prepares this transition by introducing a fn that reads the
`irreport` and `irdepth` fields and produces that `Option`. Rather than
having the `Option` carry an `u64` like `irdepth` we decided on an
`usize`, since the depth will specify the depth inside a buffer which we
will need to allocate and index on the host rather than on the tracing
target.