decoder: repurpose Payload::get_implicit_return, make it return depth
Currently, the `irreport` and `irdepth` fields defined by the
specification for `AddressInfo` and `JumpTargetIndex` payloads are
encapsulated in the `ImplicitReturn` that is only included and decoded
with the `implicit_return` feature. We plan to avoid using features for
compile time variance regarding protocol features. But rather than just
removing the `#[cfg()]` directives, we decided to replace that struct
entirely with a simple `Option` that captures the semantics of that
struct.

This means that `Payload::get_implicit_return` will not be able to
return an `ImplicitResult` in the future. And it doesn't have to, since
it is only called from contexts where we already confirmed that
`irreport` is (semantically) `true`. We thus decided to repurpose it and
make it return the `irdepth` if present.

This means that it will produce `Some` in strictly fewer cases. It also
made us rename it to `implicit_return_depth`/ Furthermote, we make it
return an `usize` rather than the `u64` currently used for `irdepth`
since we'll index a buffer on the host running the tracer.

For now it also bloats it again after we just cleaned it up. However,
that bloat will fanish as we migrate `AddressInfo` and `JumpTargetIndex`
away from `ImplicitReturn`.