tracer: rewrite Tracer::recover_ir_status
Currently, the `irreport` and `irdepth` fields defined by the
specification for `AddressInfo` and `JumpTargetIndex` payloads are
encapsulated in the `ImplicitReturn`. We plan to replace that struct
with a simpl `Option` reflecting its semantics. We thus need to
eliminate its use.

As it happends, the fn `Tracer::recover_ir_status` is partially
identical to `decoder::payload::Payload::implicit_return_depth`. Which
is not surprising since it is supposed to return `true` if `irreport` is
semantically `true`. `implicit_return_depth` returns a `Some` in exactly
those cases. We thus can express `recover_ir_status` it in a single line
using that fn.