decoder: make payload::Address hold new implicit return info
Previously, the `irreport` and `irdepth` fields defined by the
specification for `AddressInfo` payloads were encapsulated in the
`ImplicitReturn` that is only included and decoded with the
`implicit_return` feature. We plan to avoid using features for compile
time variance regarding protocol features.

But rather than just removing the `#[cfg()]` directives, we decided to
replace that struct entirely with a simple `Option` that captures the
semantics of that struct. This change makes `AddressInfo` hold such an
`Option` instead of an `ImplicitReturn`. We chose to always decode it
since it always resides at the end of a packet (supposedly for exactly
this reason). The run-time overhead should be minimal, and we can easily
just ignore the member in the tracer.