tracer: add new submodule for return stack abstractions
Rather than a proper abstraction, e.g. in the form of a struct, all data
concerning the implicit return stack is kept in multiple members of
`TraceState`, while operations on the stack are baked in feature-gated
`Tracer` fns.

For obvious reasons, we want a proper abstraction. In addition, we want
to migrate away from features for protocol related build-time variance.
Instead, we want to use traits.