tracer: redefine construction of TraceState
Rather than a proper abstraction, e.g. in the form of a struct, all data
concerning the implicit return stack is currently kept in multiple
members of `TraceState`, while operations on the stack are baked in
feature-gated `Tracer` fns. We recently introduced an abstraction called
`ReturnStack` that will replace those bits and pieces in the future.

A `ReturnStack` needs to be constructed using a piece of configuration.
Thus, it cannot be `Default`, meaning that it needs to be supplied when
constructing the `TraceState`. We therefore decided to redefine how it
is constructed, taking the opportunity to supply the instruction cache
from the outside.