tracer: add error for inability to construct a return stack
Rather than a proper abstraction, e.g. in the form of a struct, all data
concerning the implicit return stack is currently kept in multiple
members of `TraceState`, while operations on the stack are baked in
feature-gated `Tracer` fns. We recently introduced an abstraction that
will replace those bits and pieces in the future.

The currently used implementation has a known supremum for maximum
depth. Currently the check whether or not the maximum depth exceeds that
supremum is burried in one of the aforementioned feature-gated fns. With
the new implementation we'll perform that check during construction of
the `Tracer`. Since the existing error thrown by the existing check
could be used for conveying this informaiton, we feel it's not
appropriate any more.