tracer: make Tracer::new faillible
Rather than a proper abstraction, e.g. in the form of a struct, all data
concerning the implicit return stack is currently kept in multiple
members of `TraceState`, while operations on the stack are baked in
feature-gated `Tracer` fns. We recently introduced an abstraction called
`ReturnStack` that will replace those bits and pieces in the future.

A `ReturnStack` needs to be constructed using a piece of configuration,
and the construction may fail. Thus, if we want to switch over to that
new abstraction, we need to enable communication of failures from
`Tracer::new`.