tracer: migrate TraceState and Tracer to stack::ReturnStack
Rather than a proper abstraction, e.g. in the form of a struct, all data
concerning the implicit return stack is currently kept in multiple
members of `TraceState`, while operations on the stack are baked in
feature-gated `Tracer` fns.

We recently introduced a proper abstraction. This change now migrates
`TraceState` and `Tracer` to that new abstraction.