instruction: add dedicated module for Instruction retrieval
Currently, the `Tracer` will grab `Instruction`s out of `Segment`s using
its very own logic. This forces users to prepare those `Segment`s from
whatever they have. We do want both a nicer abstraction on the side of
the `Tracer` and greater flexibility for users.

This change introduces an abstraction in the form of a trait. We will
use it in the `Tracer` and allow users to supply their own
implementation.