decoder: make Synchronisation::get_branch return an Option<bool>
Previously, `Synchronization::get_branch` would return a `Result`, with
`tracer::Error` as an error type. While this may have been convenient
for users in `tracer`, it posed a dependency of `decoder` on `tracer`
that we don't want.

This change removes this dependency by making the fn in question return
an `Option`. Since we're here we make that an `Option` of a `bool`,
letting the caller do the conversion if necessary.