decoder: make Synchronization::get_privilege return an Option<Priviledge>
Previously, `Synchronization::get_privilege` would return a `Result`,
with `tracer::Error` as an error type. While this may have been
convenient for users in `tracer`, it posed a dependency of `decoder` on
`tracer` that we don't want.

This change removes this dependency by making the fn in question return
an `Option`. And sicne `Priviledge` is really just a simple enum, we
chose to return it by value rather than by reference.