tracer: only store value of sijump_p in Tracer
Previously we would store the entire `ProtocolConfiguration` in the
`Tracer`, even though we would only use a single value. This change now
stores just that value. Should we require anything else in the future,
we'll also store those values discretely, or use them for the
construction of some entity like we do with the `ReturnStack`.