tracer: eliminate Tracer::follow_execution_path
`TraceState::ir` is _meant_ to convey whether or not there is a return
stack depth in the `Payload` that is currently processed. It doesn't
make sense to drag this information through the `TraceState`, at least
not the way it's currently done. We thus want to elimate it.

This change adapts a check in `Tracer::follow_execution_path` that
previously used `Tracer::follow_execution_path_ir_state` which was
nothing but "stack depth present? -> stack depth == value"? Since we
don't have any other users for `Tracer::follow_execution_path_ir_state`,
we can remove it.