tracer: remove redundant check on IR in Tracer::implicit_return_address
`TraceState::ir` is _meant_ to convey whether or not there is a return
stack depth in the `Payload` that is currently processed. It doesn't
make sense to drag this information through the `TraceState`, at least
not the way it's currently done. We thus want to elimate it.

We used `TraceState::ir` (essentially) as a precondition for a
comparison of the `irdepth` field of the payload to the depth of the
stack held in the `TraceState`. That check implicitely already did the
right thing, making the precondition redundant.