tracer: eliminate TraceState::ir
`TraceState::ir` was _meant_ to convey whether or not there is a return
stack depth in the `Payload` that is currently processed. We previously
only use it in checks that also require the depth itself if present
anyway, and we get both pieces of information (whether or not we have a
depth and the depth) from `Payload::implicit_return_depth` anyway.

It doesn't make sense to drag this information through the `TraceState`,
at least not the way it was done previously. We thus decided to elimate
it and removed all users. This change now removes the `TraceState`
member itself.