tracer: remove feature-gating of Tracer::sequential_jump_target
We plan to avoid using features for compile time variability concerning
protocol features. In order to eliminate the `implicit_return` feature,
and as a generally sensible abstraction, we recently introduced the
`ReturnStack` trait and migrated `TraceState` and `Tracer` over to that.

The behaviour of the `Tracer::sequential_jump_target` variant for the
`implicit_return` feature already matches the behaviour for the variant
used if the feature is not active for `ReturnStack`s with a maximum
depth of zero. We thus can eliminate the latter variant.