tracer: add Builder for Tracer
Currently, `Tracer`s are constructed via `Tracer::new`. That fn takes
multiple configuration structs, all of which impl `Default`, and does a
small amount of work in order to create a `Tracer`. This is clearly one
of those cases where you want a builder.

This chace introduces a builder, but does not remove `Tracer::new`. That
will be done by a future change. We _do_ take the opportunity to use
individual values rather than `TraceConfiguration`, which allows us to
eliminate that struct in the future.