tracer: add an enum for the version of the trace specification
We currently support tracing following two (slightly incompatible)
versions of the Efficient Trace for RISC-V specification, namely the
original version 1 and the new version 2. Currently, we trace according
to version 1 when the `tracing_v1` is selected and according to version
2 if it is not selected.

We chose a feature because we did not want this to be runtime variable.
However, in hindsight the differences are miniscule and the few
additional branches we would introduce are acceptable. Thus, we decided
to switch from the feature to a runtime tracer configuration.

This change introdces an enum for the version.