tracer: add version member to Tracer
We currently support tracing following two (slightly incompatible)
versions of the Efficient Trace for RISC-V specification, namely the
original version 1 and the new version 2. We currently use a feature for
selecting versions, but we plan to migrate to a runtime configuration.

This change introduces the data mamber holding that configuration, which
is currently pinned to version 2 by the fact that we provide no means to
change it.