tracer: make version specific behaviour depend on version member
We currently support tracing following two (slightly incompatible)
versions of the Efficient Trace for RISC-V specification, namely the
original version 1 and the new version 2. Previously, the version was
selected at build time using a feature. However, we decided that the few
branches that a runtime configuration would introduce was acceptable.

This change now makes the behaviour depend on the `Tracer::version` data
member that we introduced recently.