tracer: add fn for version selection to Builder
We currently support tracing following two (slightly incompatible)
versions of the Efficient Trace for RISC-V specification, namely the
original version 1 and the new version 2. Previously, the version was
selected at build time using a feature. However, we decided that the few
branches that a runtime configuration would introduce was acceptable.

We therefore introduced a data memeber and made logic depend on that
rather than a feature. However, we previously did not provide any means
to select the version. This change now introduces a `Builder` fn for
selecting the version.