tracer: remove cache type parameter from Tracer
We plan to replace `Segment`s as a source for `Instruction`s with a
`Binary`. Since caching is not mandated by the specification (and
frankly, it doesn't make much sense anyway), we decided that, if we want
the `Tracer` to cache `Instruction`s, we would do that behind the
`Binary` abstraction.

Thus, we won't use any cache in `Tracer` itself, meaning we'll remove
the parameter at some point any way. Doing that now, before we introduce
a new parameter for the `Binary`, reduces the overall noise.