tracer: replace Segments with Binary
Previously, the `Tracer` would grab `Instruction`s out of `Segment`s
using its very own logic. This forced users to prepare those `Segment`s
from whatever they have. We wanted both a nicer abstraction on the side
of the `Tracer` and greater flexibility for users.

We introduced the `Binary` trait as an abstraction that could also be
easily implemented by users of this library, or by ourselves for third
party types. This change now migrates `Tracer` from `Segment`s to a
`Binary`. Sadly, this is a big change that we could not reasonably
devide into smaller chunks.