tracer: remove instr_cache from TraceState
Previously, `Tracer` would cache `Instruction`s it retreived using an
`InstructionCache`. Caching of `Instruciton` is not mandated by the
specification and it does generally make little sense to do this within
the `Tracer`. We thus decided to remove `InstructionCache`s.

This change removes a leftover of the recent migration from `Segment`s
to a `Binary` as an `Instruction` source.