tracer: make Error report source for CannotGetInstruction
The `Builder` used by a `Tracer` will emit an error of some type if the
retrieval of an `Instruction` failed. We propagate this error upwards in
an `Error::CannotGetInstruction`.

This change now also makes `Error` communicate this properly as a
source. This means we now impl `core::error::Error` for `Error` only
depending on traits implemented by `Builder::Error`, but that's
something we can (probably) accept.