instruction: remove Segment
In the past, `Segment`s were used as an `Instruction` source in
`Tracer`. However, we since switched to the `Binary` trait which is a
better abstraction.

We could impl `Binary` for `Segment`, or a slice of them, but the
benefit that type provides isn't that big. For now, we just ditch it
completely.