decoder: add a builder for Decoders
Currently, `Decoder`s are build via `Default::default` or `Decoder::new`
with a `config::Protocol`. After that, they need to be modified to hold
the actual data they decode. Which consumes self since it may alter the
lifetime of the decoder. This means we already have something similar to
a builder.

We decided to go all the way and introduce a proper builder. We keep the
old methods for creating a `Decoder` usable until we migrated all users.