decoder: remove means to construct Decoder without Builder
Previously, `Decoder`s were build via its `Default` impl or
`Decoder::new` with a `config::Protocol`. After that, they needed to be
modified to hold the actual data they decode. Since that scheme was
already close to a proper builder we decided to introduce one and
migrated all users over to the new `Builder`. This change now removes
the old fns for constructing a `Decoder`.