decoder: add module for branch map utilities
There are multiple packet formats containing a branch map. In the
structs carrying the decoded data, they were represented by one variable
each for the branch count and the contents. We want some form of
encapsulation for that. And ideally, we want to use the same abstraction
inside `tracer`.

This change introduces such an encapsulation in the form of a struct.
Rather than the `u32` we used for branch maps in the past, we use an
`u64` for a bit more head space. This may be unnecessary in practice,
but we would need to study the tracing logic in detail to know that.