decoder: add helper for decoding branch counts
Naturally, we want to decode `branch::Map`s. However, branch maps come
in different variants. While for `JumpTargetIndex`, a `branches` field
with the value `0` simply indicates an empty branch map, it indicates a
"full" branch map for `Branch` (as well as the absence of an address).
We thus cannot simply impl `Decode` for `branch::Map`.

This change introduces a helper that we can use to at least abstract the
reading of a branch map a little bit.