tracer: make TraceState::branch_map a decoder::branch::Map
Previously, we would handle branch maps as a branch count and a separate
value containing the actual branch information in both the `decoder` and
`tracer`. We recently introduced a proper encapsulation in `decoder`
which was specifically designed to be usable in the `tracer`. This
change now migrates `TraceState` and `Tracer` over to that utility.