tracer: migrate Tracer to AddressMode
We chose to express the address mode via a proper enum. This change
migrates the `Tracer` itself. We also take the opportunity to simplify
the computation of new values for `TraceState::address` inside
`Tracer::_process_te_inst`. We chose to preserve the existing behaviour,
which is likely broken for delta mode. We deem fixing that out of scope
for this change.