tracer: extract address info in Tracer::follow_execution_path
We plan to narrow down what data `TraceState` holds. In particular, we
want to eliminate everything that can be trivially extracted from the
current `Payload` since we assume that to be availible. Both
`TraceState::notify` and `TraceState::updiscon` fall into this category.

This change prepares their elimination by moving their extraction to
where they are actually needed. This makes that particular logic a bit
more convoluted, but we plan to clean that up in the future anyway.