tracer: remove Tracer::recover_status_fields
We plan to narrow down what data `TraceState` holds. In particular, we
want to eliminate everything that can be trivially extracted from the
current `Payload` since we assume that to be availible.

Both `TraceState::notify` and `TraceState::updiscon` fall into this
category. We thus eliminated their use, and all calls to
`Tracer::recover_status_fields`, which does nothing but set those.