tracer: move TraceState to dedicated submodule
The `tracer` module houses the tracing algorithm, which is mostly a
translation of pseudo-code from the Efficient Trace for RISC-V document.
The result is not particularly ideomatic Rust.

We plan to switch form the existing collection of reporting callbacks to
an `Iterator` as the primary reporting interface in the future. For
that, we would prefer the logic for advancing the state from one
instruction to another more encapsulated. We ideally want to adjust the
`TraceState` when receiving a payload and then pull PCs from that state
until we hit some stop condition. We also want the state to be more
opaque.

This change prepares this work by moving `TraceState` to a new submodule
under the name `State` as is. Future changes will add fns that will
ultimately replace parts of the old `Tracer`.